Fix rebasing errors
