+++ tools/CMakeLists.txt
@@ -40,24 +40,9 @@ mirtk_add_executable(
 )
 
 mirtk_add_executable(
-  resample-to-iso-grid
+  centre-volume
   SOURCES
-    resample-to-iso-grid.cc
-  DEPENDS
-    LibCommon
-    LibNumerics
-    LibImage
-    LibIO
-    LibRegistration
-    LibTransformation
-    LibSVRTK
-    ${TBB}
-)
-
-mirtk_add_executable(
-  centre_volume
-  SOURCES
-    centre_volume.cc
+    centre-volume.cc
   DEPENDS
     LibCommon
     LibNumerics
@@ -100,7 +85,7 @@ mirtk_add_executable(
 )
 
 mirtk_add_executable(
-  register_landmarks
+  register-landmarks
   SOURCES
     register-landmarks.cc
   DEPENDS
